DAMD/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "damd"
version = "0.1.0"
description = "Data As Metadata in Data - Embed small related data within larger host files"
authors = [{name = "DAMD Project", email = "dev@damd.io"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "msgpack>=1.0.0",
    "zstandard>=0.19.0",
    "cryptography>=3.4.0",
    "pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.991",
]

[project.scripts]
damdcat = "damd.cli:damdcat"
damdls = "damd.cli:damdls"
damdrm = "damd.cli:damdrm"
damdinfo = "damd.cli:damdinfo"
damdwrite = "damd.cli:damdwrite"

[project.urls]
Homepage = "https://github.com/damd-project/damd"
Repository = "https://github.com/damd-project/damd"
Documentation = "https://damd.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["damd*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
